preprocess and genePairPlotting
preprocess - Reads in GEM and creates a condensed version of the data.  This should be the last functional version.  I was working on updates, but they are not complete yet.
genePairPlotting - Plots the average of the first and second genes for every cluster; colors the points based on cancer tissue type